🧹 Moved 'dummy_result' fixture to 'glotaran/project/test/test_result.py'
That way it can be imported and by tests and there is less code duplication.
